fastapi_OAuth2/pyproject.toml
[tool.poetry]
name = "fastapi-oauth2"
version = "0.1.0"
description = ""
authors = ["Muhammad Saad <saadjameel2001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.30.1"
python-jose = "^3.3.0"
fastapi = "^0.111.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
